fix: accountBalance, transactionAmount 자료형 수정
transactionAmount 자료형 수정(Integer -> Long)
transactionAmount 자료형 수정(Long -> Integer)

Related to: S12P21A208-27

+++ blind_fintech/src/main/java/com/blindfintech/domain/transction/controller/request/TransactionRequest.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 public class TransactionRequest {
     private Integer senderAccountId;
     private Integer receiverAccountId;
-    private Long transactionAmount;
+    private Integer transactionAmount;
     private String transactionName;
 }
